move switch related constants
